CogVM source as per VMMaker.oscog-eem.1877
Cogit:
Revise Clement's two-path code for immutability to allow it to be used for
frameless setters containing more than one inst var store, avoiding multiple
store checks by checking once for the receiver being young.  Good for a 4%
speedup in the binary tree benchmark.

Rename needsTwoPath to useTwoPaths.  Refactor StackToRegisterMappingCogit>>
scanMethod so that it is used by RegisterMappingCogit and SistaCogit without
having to duplicate an already large method with very minor varations 3 times.

Add an isInstVarStore property to CogBytecodeDescriptor and set it in the
bytecode tables.

Fix bug in ARM's concretizeSqrtRd.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3734 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: ac654218f3a3bd6d07cbffed9e4e03dccb80cb71