CogVM source as per VMMaker.oscog-eem.1883
On Spur, handle external primitive failures answering PrimErrNoMemory by doing
a scavenge and retrying and then on subsequent PrimErrNoMemory failure a full
GC and then retrying.  Rename maybeRetryFailureDueToForwarding to
maybeRetryPrimitiveOnFailure.

In the Cogit, change the retry mechanism so that the primitive call is made in
the run-time, instead of the run-time answering whether the primitive should
be retried and returning to machine code that then retries.

Discussion:
First, this mechanism applies only to external (plugin) primitives.  Spur's core
primtives are written in the expectation that image-level primitive failure code
will deal with the failure and retry.  This corresponds to the VisualWorks style
and allows the image to implement GC policy (e.g. to favour reclamation over
growth).  So with the current Spur code, heavily influenced by the author's
experience with VisualWorks, it would be inapproprate to retry core primitives
on being out-of-memory.  The pathology is that attempts at very large
allocations which will inevitably fail will cause a GC in the VM, very possibly
followed by a GC from the image level primitive failure code.  Whereas leaving
it to the image, the image code may be intellkigent enough to spot an allocation
too large to ever succeed and avoid the GC altogether.

I'm not sure if I like this auto-retry or not.  Perhaps it is a good idea and
the image-level primitive failure code should be changed; there would certainly
be fewer image-level methods.  But that means fewer hooks for the image to
implement GC policy.  retryPrimitiveOnFailure which manages the retry isn't
exactly simple.

Slang:
Fix a bad bug in type inferrence as exemplified by primitiveTimesTwoPower and
its arg temporary.  A variable cannot be typed in the current pass if it is
first assigned the result of an as-yet-to-be-typed send.  If it is assigned the
result of a subsequent send then the first assignment will be ignored, possibly
leading to a badly inferred type.  For example in primitiveTimesTwoPower arg is
assigned a few times:
	arg := self stackTop.
...
	arg := objectMemory integerValueOf: arg.
...
	arg := twiceMaxExponent negated
...
	arg := twiceMaxExponent
and the type of twiceMaxExponent is #int.  If arg's type is inferred from
twiceMaxExponent because stackTop & integerValueOf: are as yet untyped, then it
will be inferred to be #int.  Whereas it should be defined as #sqInt from both
stackTop & integerValueOf:, since merging #int & #sqInt yields #sqInt.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3738 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 6edb7ee88e494863f05adad1b79b108a7e98e5d7