VMMaker.oscog-eem.1884 Fix multiple bytecode set selection in the Spur 64-bit VM, which must use signedIntFromLong64 to test for a negative header.
In 1884 the Fabian Society is founded in London, Georges-Pierre Seurat paints Bathers at Asnières, and the eight-hour workday is first proclaimed by the Federation of Organized Trades and Labor Unions in the United States. This date, called May Day or Labour Day, becomes a holiday recognized in almost every industrialized country.