CMakeVMMakerSqueak-tty.128.mcz Writing a terse guide as I build a  Linux64x64SqueakCogSpurConfig from scratch
Started a refactoring sweep .
Refactored compilerFlags, compilerDefinitions linker flags methods to use the redirect pattern. 

wrote tests to check new methods

consider this a stage save. I have more to do on this line of attack

Left a note to myself to add a method protocol named "customizable" or something to cue the end-user that these methods should be customized by them.

I expect the number of redirect method constructs to drastically reduce and the number of redirect methods that are directly customizable to increase.