CogVM source as per VMMaker.oscog-eem.1890
Fix regression in Sista introduced in VMMaker.oscog-eem.1877.  The IMMUTABILITY
scanMethod:  should set useTwoPaths only for methods that would be frameless
were it not for a store (which needs a frame if immutable).  The regression
was to set useTwoPaths in frameful methods containing a store; a stupid slip
by your's truly.