DSAPlugin as per CryptographyPlugins-eem.12
Fix DSAPlugin for 64-bits; leftRotate:by: needs to be declared as returning
unsigned int, and in simulaiton needs to keep the result poegged to the 32-bit
range.  Improve performance by simplifying a number of argument validation
sequences.  fetchClassOf: and isWords: isBytes: all check for immediates, so
there is no need to use stackObjectValue:, and using the and:and:and: form
instead of the success: foo. success: bar. form leads to fewer calls.