VMMaker.oscog-eem.1893 Fix a couple of cases that should use mositiveMachineIntegerValue: and nuke an unnecessary stackObjectValue:.
In 1893: Riots of Mons during the Belgian general strike of 1893, The day after, Belgian parliament approved Universal suffrage. The Independent Labour Party of the United Kingdom has its first meeting. The United States overthrows the Kingdom of Hawaii: Lorrin A. Thurston and the Citizen's Committee of Public Safety in Hawaii with the intervention of the United States Marine Corps overthrow the government of Queen Liliuokalani.