CogVM source as per VMMaker.oscog-eem.1891
Simplify parameter validation in several plugins, especially FloatArrayPlugin.
stackObjectValue: checks for an object, failing if it accesses an immediate.
But so do isWords:, isBytes: et al.  There is no need to check twice.
Further, the
	interpreterProxy success: expr. ... interpreterProxy failed ifTrue: [^nil]
form contauns more calls then the
	expr ifFalse: [^interpreterProxy primtiveFail]
form and so is slower.

Trying the Travis infrastructure to test this for the first time.  Gulp.  But I'd love others to eyeball this change (and use the style in future).