VMMaker.oscog-eem.1898.mcz Fix bad bug with adoptInstance: for contexts (e.g. .| i | Array adoptInstance: (i := thisContext sender sender shallowCopy). i!).  Slots beyond stack pointer must be nilled.  Further fix bug in externalInstVar:ofContext: for pc access.  Add sufficient assert machinery to allow stack pointer to be determined in external regime (framePointer valid). Add intAt: support for 64-bit simulation (cf VMMaker.oscog-cb.1895).
In January of 1898 novelist Émile Zola's open letter to the President of the French Republic on the Dreyfus affair, J'accuse…!, is published on the front page of the Paris daily newspaper L'Aurore, accusing the government of wrongfully imprisoning Alfred Dreyfus and of antisemitism. In February of that year the electric car belonging to Henry Lindfield of Brighton runs away on a hill in Purley, London, England, and hits a tree; thus he becomes the world's first fatality from an automobile accident on the public highway, and the USS Maine explodes and sinks in Havana harbor, Cuba, for reasons never fully established, killing 266 men. This event helps lead the United States to declare war on Spain.