Eliminate some -Wabsolute-value compiler warnings
sizeField() returns a sqInt.
sqInt is 64bits long on 64bits images.
Taking abs(sqInt) will truncate the size to 32bits.

That's probably benign because the size will rarely exceed 2^31, so we effectively truncate nothing.
But we have a correct way to take absolute value of sqInt (SQABS) and long (labs) so let's just use them and eliminate the false positive warnings.