Make positiveMachineIntegerValue prototype LLP64 compatible.
Apply similar change on signed version.
Most uses are for returning a machine word, aka a usqIntptr_t,, a few for getting a SmallInteger value in Spur (where usqInt matches usqIntptr_t)
This change shall be applied on VMMaker side in order to be fully LLP64 compatible.
It is compatible with existing configurations where usqIntptr_t is defined as unsigned long.

Reminder: we do not use usqInt for the aim of running 32bits image on 64bits VM.