Fix the printf format used in platforms/win32
Note: the problem is that we do not know if sqInt is int, long or long long.
Same for sqIntptr_t which can be long or long long (on LLP64 like windows x86_64)

So we do not know which format to use, %d %ld or %lld in printf function...
The standard (C99) solution is to use macros for printf format (and scanf format)
See for example http://en.cppreference.com/w/c/types/integer
or http://stackoverflow.com/questions/1255099/whats-the-proper-use-of-printf-to-display-pointers-padded-with-0s

Since we define our own types sqInt sqIntptr_t, we also define our own macros PRIxSQINT PRIxSQPTR etc...
We thus abide to the standard way of doing things...

Hopefully, these changes should eliminate all the -Wformat warning encountered on platforms/win32 in a LLP64 compatible way.
(until we declare the prototype of some implicitely typed functions which might change the deal and uncover more problems).

Similar changes will have to be applied to Smalltalk (slang) code (VMMaker.oscog branch).