Fix incorrect use of recordMouseEvent  in sqHostWindowPlugin
This function was used with implcit declaration...
Unfortunately, the implementation (in sqWin32Window.c) does not fit the usage...
it takes a second parameter, number of clicks.
I suspect the plugin does not work for a long time...

The corection is a rough work around, not taking into account multiple clicks, and/or motions during click.