Declare prototypes of functions used in platforms/win32
Not doing so is dangerous because we might pass an int (32 bits) instead of an sqInt (64 bits)
It is thus sane to eliminate the warnings of the kind:
../../platforms/win32/vm/xxx.c:58:4: warning: implicit declaration of function 'yyy' is invalid in C99 [-Wimplicit-function-declaration]

TODO:
Current workaround leads to more duplication of prototypes which is not satisfactory.
We should better use header files in the future.