Fix interpreterProxy prototype for positive32BitIntegerFor()
The parameter has been switched to 'unsigned int' for a while in VMMaker.oscog branch.
Declaring sqInt is not compatible with 64 bits image plugins, and generates a C Compiler warning.

I suspect the warning to be false positive, because we copy function address, and the one we use has correct prototype.
Correct  prototype definition will avoid relying on such hypothesis :)