Fix missing maxOldSpaceSize again for SpurStack VM
Note that the variable is declared usqInt, not unsigned long as current version of cointerp.c
This is because it's the same thing on almost every machine but win64 (LLP64)
So it will just work but on win64. But win64 does not work anyway.
In Win64, it's necessary to amend cointerp.c and declare maxOldSpaceSize as usqInt.
That's a VMMaker pending change. We take a little advance for when it will be ready.

Note that the problem first originated in https://github.com/OpenSmalltalk/opensmalltalk-vm/commit/82d1c33a1c5756b0a76bd37539d15a4824368d59
which was too big to review carefully. Please use more atomic changes next time.
Combinated with false green from appveyor, that's been tricky.