Merge VMMaker.oscog-cb.1912, VMMaker.oscog-nice.1911
Use same FileDirectory tricks as Cog* for Stack* generation
This way we can generate sources on windows (with recent trunk corrections for pathFromURI:).

Don't put a trailing / to '../'. in VMMaker initialization.
Indeed, some FileDirectory hacks work only for '..'
For example, (FileDirectory on: '..') is ok, (FileDirectory on: '../') is not.

Correct the prototype of genDoubleFailIfZeroArgRcvr:arg:
It's used as a function taking int, not #sqInt
See #genDoubleArithmetic:preOpCheck: and #genDoubleArithmetic:preOpCheck:boxed: for 64bits