VMMaker.oscog-eem.1915 Spur: Implement machine-code shallowCopy for Spur, excepting 0 slot objects, contexts, methods and objects with overflow size.
Fix the commentary for Xbr, X16r & X32r addressing modes.

Make sure that markAndTrace: is never inlined.

N.B. The 64-bit Cogit appears to be broken (an MNU early on, followed by a 0 arg to value: later in trying to report the MNU).  No time to fix it now.  When did it break I wonder?