CogVM source as per VMMaker.oscog-eem.1930
Cogit:
Implement compacting CPICs on method zone reclamation rather than freeing CPICs
that have one or more freed entries.  This for more stable analysis in the
Sista VM.
Make MaxCPICCases a constant (class var) rather thnan a variable.

Cogit SistaV1/FullBlocks:

When a loop is taken enough time to trigger jit compilation, it needs in the
SistaV1 bytecode set to check if the activation is a full block and call the
correct JIT compiler method if this is the case.

SistaV1 bytecode set needs to define its primitive error code bytecode
(long temporary store bytecode)

Now that the trampoline for context inst var store can be called on single
contexts, we need to update the trampoline to work with single contexts

Comment the register saving convention for  the context inst var access
trampolines in the JIT.

Added immutability check in maybe context object store in jitted code.

Added an annotation for inst var store in the SistaV1 bytecode set to tip the
JIT if the mutated object may be a context or not.

StackSpur VM are now generated with dual bytecode set support.

Quick fix on the push nil functions for the sista bytecode set.

Added support from error code while scanning block bodies.

VMMaker.oscog-cb.1920
This commit makes a difference between the usage of the sistaV1 bytecode set and
the Sista VM. It is now possible to compile the VM with the SistaV1 bytecode set
but without the SistaVM option. This way, one can use and stabilize:
- directed super sends
- full block closures
without using unsafe operations and traps. I believe we'll migrate to the sista
runtime in multiple steps, and the first step is to use those features and have
images compatible with those features without runtime optimizations.

The SistaV1 bytecode set allows to use directed super send and full block
closures. Enabling the bytecode set enables those features.
(SistaV1BytecodeSet option)

The SistaVM option, combined with SistaCogit, enables the profiling counters,
the unsafe operations and traps. If disabled, the bytecodes from the SistaV1
bytecode set corresponding to traps or unsafe operations are considered unknown
bytecodes.

All unsafe operations and traps related JIT code is now in SistaCogit.
Those bytecodes are considered as unknown bytecodes in superclasses.

I've also changed some flag logic, as the flag marking optimized method is
bit 16 (previously upper literal bit) and not the flag bit.