VMMaker.oscog-eem.1937.mcz More full block support. Find the home of a full block for context printing. Fix opc mapping for full block closures (mustMapMachineCodePC:context:).
In the Cogit change inBlock to be an integer value, 0 for methods, 1 for vannilla blocks and 2 for full blocks (to reuce the amount of refactoring now and when removing vanilla block support some time in the future).

Add trampolines for creating small and large full block contexts in the Spur object representation.

Get frameless full block prolog generation more correct than it was ;-)