CogVM source as per VMMaker.oscog-eem.1948
General:
Fix off-by-one error in an assert

Fixes related to Immutability / read-only objects:
- weak structures can't be mmutable
- primitiveStringReplace fails if the receier is immutable only if it attempts to mutate something.

SistaBytecodeSet/FullBlocks:
Re-enabled full block compilation to machine code.

Fix marryFrame:SP:copyTemps: for full blocks; the closureOrNil field must
still be set.  Add an assert to primitiveFindNextUnwindContext that caught a
side-effect of the bug.

Added assertion ensuring that when a method is jitted, the compiled code
jitted is a compiled method and when a block is jitted, the compiled code
jitted is a compiled block.

Improve stack backtrace printing for full blocks.

Disabled 2 paths compiltion on full blocks.

Find the home of a full block for context printing. Fix pc mapping for full
block closures (mustMapMachineCodePC:context:).

In the Cogit change inBlock to be an integer value, 0 for methods, 1 for
vannilla blocks and 2 for full blocks (to reuce the amount of refactoring
now and when removing vanilla block support some time in the future).

Add trampolines for creating small and large full block contexts in the Spur
object representation.

Get frameless full block prolog generation more correct than it was ;-)

Added a deny to ensure the block can be frameless. Works both for blocks and
full blocks.