CogVM source as per VMMaker.oscog-eem.1952 (ignoring Lowcode changes for now).
Make VMProfileMacSupportPlugin support 64bits (at least for collecting modules).

Make the minimum supported version on Mac OS X in the Info.plists 10.9.  It
actually needs to be computed from the SDK used to build the product, so this is
a temporary fix.

Eliminate a cpp warning in platforms/Cross/vm/sqVirtualMachine.h with Spur VMs.