CogVM source as per VMMaker.oscog-eem.1968
Fix initialization of Spur's newSpace.  The old code was convoluted and wrong.

Reduce the default number of stack pages in Spur to 50 to reflect GC tuning
work at Cadence.

Add vmParameter 19 to answer the value of the scavengeThreshold, which is the
effective size of eden. (Spur only).

Update the Mac HowToBuilds to describe solution to Sierra build failure due to
missing libssl and libcrypto.