Fix bug https://github.com/OpenSmalltalk/opensmalltalk-vm/issues/23 ioGetClockLogSizeUsecsIdxMsecsIdx prototype is not 64bits compatible
The problem is in primitiveClockLogAddresses for which type inference for #runInNOut result into an int.
But ioGetClockLogSizeUsecsIdxMsecsIdx expect a pointer on a sqInt, not a pointer on an int as first parameter.
Workaround by explicitely decalring #runInNOut as sqInt.

The fix is in oscogLLP64 branch (http://smalltalkhub.com/mc/nice/NiceVMExperiments/main) since VMMaker.oscogLLP64-nice.1915, Time: 29 July 2016, 12:36:23.284336 am