CogVM source as per VMMaker.oscog-eem.1975
Fix the root cause of the bugs wrngly fixed in VMMaker.oscog-eem.1971 & 1972.
Eliminate the extBFirstZero variable and its effects on assembling a signed
extB.  This restores the Newspeak and no-access-control Cadence Newspeak VMs.

Eliminate C compiler warning about taking absolute value of positive integer:
correctly store the value in a signed integer, then take the absolute value,
in #rewriteCPICJumpAt:target: and #findEmptySegNearestInSizeTo:

Fix a typo in #flushICacheFrom:to:.

Use #sqLong type in is32BitSignedImmediate: because it's guaranteed to be
64bits long. #long can be 32bits for LLP64 compilers.

Another 64bits fix for MacMenuBarPlugin.

Fix bug https://github.com/OpenSmalltalk/opensmalltalk-vm/issues/23
ioGetClockLogSizeUsecsIdxMsecsIdx prototype is not 64bits compatible

The problem is in primitiveClockLogAddresses for which type inference for
a pointer on a sqInt, not a pointer on an int as first parameter.
Workaround by explicitly declaring #runInNOut as sqInt.