VMMaker.oscog-nice.1979 Introduce sqIntptr_t and usqIntptr_t in VMMaker (image side) for LLP64 compatibility - but do not yet use them. These are integers long enough to hold a target machine pointer (BytesPerWord). These should later replace most usage of #long and #'unsigned long'. They are already defined in sqMemoryAccess.h.
Add enough machinery for generating and simulating these types, inferring size, converting signedness etc....

Introduce asUnsignedIntegerPtr which will convert to (cast to) a usqIntptr_t.
same for asIntegerPtr -> sqIntptr_t.
They should replace most usage of asUnsignedLong and asLong in the future.
Unless we don't want a machine word, but an oop, in which case we should use asUnsignedInteger (-> usqInt) or asInteger (-> sqInt) (typically to turn a longAt: into an unsignedLongAt: - because longAt: answers a sqInt not a long!)

Restore the size of sqInt/usqInt as BytePerOop.
BytesPerOop matches BytesPerWord for Spur currently, but the CCodeGenerator should be Spur-agnostic if possible.

Move sqInt/usqInt check ahead in simulated sizeof: because most variables are of this type.