VMMaker.oscog-nice.1982.mcz Replace remaining long/unsigned long with sqIntptr_t/usqIntptr_t. Replace asLong/asUnsignedLong senders with asIntegerPtr/asUnsignedIntegerPtr. This should be essentially for those int containing a target machine pointer/address.
For some printf cCode (used for debugging), the %ld does not work in LLP64.
We use the standard C way for compatible 32 & 64bits code:
use PRI_format_Type macros as the format (mainly PRIxSQPTR).

This way, VMMaker is now LLP64 friendly.
For other conventions (ILP32 and LP64), this does not change anything since sqIntptr_t is equivalent to long.