RegisterAllocatingCogit Get the 64-bit system as far as the isInMemory send in SmalltalkImage>>#send:toClassesNamedIn:with:.
Main change is to not merge following a return (there is no mergeSimStack at a fixup).

Fix several asserts checking the spill state of the stack.
Reorder the compilationTrace flags and add tracing of compilation in cog:selector: and cogFullBlockMethod:numCopied:.
Improve printing of constasnts in the sim stack entry classes.