RegisterAllocatingCogit Get the 64-bit system as far as the second \\ send in IdentityDictionary>>#scanFor:.
Set a simStackEntry's liveRegister in popToReg: and storeToReg: in addition to in genStorePop:TemporaryVariable:.  When setting liveRegister: in this way, look through the simStack for any and all copies of the entry having its liveRegister set, and set their liveRegister to match.

When ensuring that ensureReceiverResultRegContainsSelf, nuke any liveRegister in non-self stack entries that is equal ReceiverResultReg.
 
Make inlined genSpecialSelectorArithmetic keep register values live around the uncommon send by cloning the simStack after generating inline code and reconciling the register state after the send.

liveRegister must be taken into account when asking for a simStackEntry's registerMask to arrange that flushing/spilling occurs.

Print the optStatus as well as the simStack when tracing the simStack.