RegisterAllocatingCogit & SistaCogit
Refactor StackToRegisterMappingCogit>>compileBlockBodies to eliminate SistaCogit>>compileBlockBodies (now stale because it was missing the literalsManager save/resetForBlockCompile sends)).

Refactor simStack printing to allow CogRASSBytecodeFixup>>printStateOn: to print its simStack on an arbitrary stream.

The only change in semantics:
I *think* RegisterAllocatingCogit>>ensureFixupAt: needs to set the mergeSimStack of a fixup if the fixup needsFixup and doesn't have a mergeSimStack.