RegisterAllocatingCogit: methodOrBlockNumTemps must not be reset to the actual value ignoring the initial pushNils.  Doing so wrecks spilling.
scanMethod must add an extra fixup for conditional branches since, if preceded by a special selector send, there may be two fixups created, one for each target, the target of the jump and the fall-through after the non-inlined send (see e.g. the two ensureFixupAt: sends in genBinaryInlineComparison:opFalse:destReg:).

This gets execution as far as the send of #<= in the loop in Interval>>#collect: which ends up passing nil as the second argument.  It maybe time to tackle inlining of special selector comparisons.