VMMaker.oscog-nice.2001 Correctly marshal the trampoline arguments for Win64 x64 ABI. The registers for passing first 4 integer/pointer parameters are RCX, RDX, R8, R9.
More over, it's necessary to reserve shadow space on the stack so that the callee can save those 4 registers even when there is less than 4.
Consequently, it's mandatory for cogit to invoke genRemoveNArgsFromStack: even when numArgs = 0, and let the backEnd decide what to do.

Correct a typo in cResultRegister comment.