VMMaker.oscog-eem.2021.mcz RegisterAllocatingCogit: Rename moveSimStackConstantsToRegisters to moveVolatileSimStackEntriesToRegisters sicne we also have to capture inst var refs (anything not relative to FPReg).
Fix slip in reconcileForwardsWith:.

ensureNonMergeFixupAt: shoudl only set the mergeSimStack of the fixup if it doesn't yet have one.

When merging we have tio make sure that simSelf's liveRegister is up-to-date, not just optStatus.  Oh how nice it would be to only have simSelf and be able to nuke optStatus. [the issue is having two defs for a simStackEntry in StackToRegisterMappingCogit, one with liveRegister, needed for simSelf, one not, for normal entries].

Get fixups to remember the bytecode pc that set them, and print the bcpc(s).  Add printing simSelf to traceSimStack.

Execution now gets to Process>>popTo: during SmalltalkImage>installLowSpaceWatcher, 187 methods compiled and executed.