RegisterAllocatingCogit: Abstract away voiding the optStatus into voidReceiverOptStatus so that RegisterAllocatingCogit can override to clear the liveRegister on simSelf.
Add voidReceiverResultRegContainsSelf that scans the stack for occurrences of simSelf in a register on stack and to either flush (StackToRegisterMappingCogit) or void liveRegister (RegisterAllocatingCogit).  Except that I suspect RegisterAllocatingCogit may have to flush in certain curcumstances too.

Add asserts to check that optStatus and simSelf are in sync w.r.t. ReceiverResultReg.

Fix a bad slip in existsInstVarRefBeforeSendOrReturn.

Nuke an obsolete NewspeakV3 bytecode.

Very close to working properly.  System no longer crashes in JIT or assert failure but runs, albeit looping on an MNU.