RegisterAllocatingCogit:
Fix a slip in reconcileForwardsWith:; if a constant is merged with something other than the same constant then the resulting stack entry cannot be constant.  This fixes the bug compiling the or: in UnixFileDirectory>>#setPathName:.

Refactor mergeWithFixupIfRequired: & mergeCurrentSimStackWith: to pass fixup so that mergeCurrentSimStackWith: can update isReceiverResultRegSelf correctly.  This fixes the bug accessing array for the second send of #size in WeakSet>>scanFor:.