VMMaker.oscog-eem.2029 Fix occasional start-up crashes on linux x64.  The choice of 16rBADA550 for a constant in closed PIC prototype generation (used to compute offsets in PICs) occasionally coincides with the address for code chosen by address space randomization, and so confuses the Cogit into generating pc-relative addressing to create this address, hence causing mayhem later on when the wrong sizes for instructions are used to decode closed PIC contents.  Hence also beef up the asserts regarding distinguishing constants planted in code on X64 for MoveCwR PushCw and ArithCwR.
Implement the Cogit side of mustBeBoolean processing in the RegisterAllocatingCogit.

Add a variable to help debug the crashes in callbacks experienced in the Pharo VM.  The variable is a set of bit flags tracing the path through returnAs:ThroughCallback:Context:.

Fix a compilation warning with the definition of sigsetjump & siglongjmp in the interpreter.