CogVM source as per VMMaker.oscog-eem.2030
Interpreter:
Improve debugging of stack imbalances on primitive invocation by restoring the
stack pointer after calling failUnbalancedPrimitive.

Add a variable to help debug the crashes in callbacks due to stack imbalance
experienced in the Pharo VM.  The variable is a set of bit flags tracing the
path through returnAs:ThroughCallback:Context:.

Fix a compilation warning with the definition of sigsetjump & siglongjmp in the
interpreter.

Do to signed32BitValueOf: what has been done to signed32BitIntegerFor:, i.e.
only invoke noInlineSigned32BitValueGutsOf: for the case of LargeInteger
deciphering.  When the oop is a 31bits SmallInteger, don't bother entering
such function - just inline.

Cogit:
Fix occasional start-up crashes on linux x64.  The choice of 16rBADA550 for a
constant in closed PIC prototype generation (used to compute offsets in PICs)
occasionally coincides with the address for code chosen by address space
randomization, and so confuses the Cogit into generating pc-relative addressing
to create this address, hence causing mayhem later on when the wrong sizes for
instructions are used to decode closed PIC contents.  Hence also beef up the
asserts regarding distinguishing constants planted in code on X64 for
MoveCwR PushCw and ArithCwR.

Fix the order of generating fixups and pushing the dummyValue when no deadCode
in genSpecialSelectorEqalsEquals:.  This didn't matter with the
StackToRegisterMappingCogit but in the RegisterAllocatingCogit now that teh
stack is not flushed on conditional branch, it is critical; the fixup must be
established before the dummyValue is pushed.

Fix a slip in genSubConstant:R: which always used TempReg and didn't use its
argument.

Abstract away voiding the optStatus into voidReceiverOptStatus so that
RegisterAllocatingCogit can override to clear the liveRegister on simSelf.

Add voidReceiverResultRegContainsSelf that scans the stack for occurrences of
simSelf in a register on stack and to either flush (StackToRegisterMappingCogit)
or void liveRegister (RegisterAllocatingCogit).

Improve Spur genGetClassObjectOf:into:scratchReg:instRegIsReceiver: to avoid
reg-reg copies.

Make #~~ a special selector instead of #blockCopy:.  If the bytecode compiler
compiles #~~ using the special selector bytecode, then the pointer comparison
with #~~ is inlined as it is with #==.  This leads #~~ to have the same
performance as #==. It also means #~~ is not a primitive call, as #==.  To
integrate that in the Squeak / Pharo image, one needs to add #~~ instead of
if not already the case.

SistaCogit:
Added a setting so that subclasses of cogit can choose to refuse early openPIC
promotion.  This leads to much, much, *much* better type information on PICs in
the Sista VM. No changes in the normal VM.

Plugins/Primitives:
Fix an unused value warning in Balloon/Bezier.  I (NC) think that the intention
was to use 2nd derivative to decide direction when 1st derivative is null.

Fix primitiveBeCursor - generated code could dereference a null pointer
offsetObj due to misplaced block end.  Also a null pointer cursorObj could be
dereferenced if 2 arguments or more were passed to the primitive.

Make pharo.stack.spur's plugins agree with pharo.cog.spur's

git:
Ignore zip files in image directory.